+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+  * Polymorphic value bindings now work properly with module type
+    ascription.
+
   * The type checker no longer requires types used inside local
     functions to be unambiguous at the point where the local function
     is defined.  They must still be unambiguous by the time the